fix INT_MIN envaluation to -0

+++ main.h
@@ -32,7 +32,7 @@ int _print_bin(va_list);
 
 
 /* print recursively */
-int _rec_print_num(long int); 
+int _rec_print_num(int); 
 int _rec_print_str(char *);
 int _rec_print_bin(unsigned int);
